requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
langgraph==0.3.5
psycopg2-binary==2.9.10
pgvector==0.3.6
python-dotenv==1.0.1
openai==1.58.1
pydantic-settings==2.7.0
docker==7.1.0